+++ ZF/ZFCore/zfsrc/ZFCore/ZFCoreDef/ZFCoreLog.cpp
@@ -11,6 +11,9 @@
 #include "ZFCoreSPrintf.h"
 #include "ZFCoreStaticInitializer.h"
 #include "ZFCoreArray.h"
+        #include <QDebug>
+        #define zfzfzfLog(fmt, ...) \
+            qDebug(fmt, ##__VA_ARGS__)
 
 ZF_NAMESPACE_GLOBAL_BEGIN
 
@@ -78,6 +81,7 @@ void _ZFP_zfCoreLogV(ZF_IN const ZFCallerInfo &callerInfo,
     }
 
     _ZFP_ZFCoreLogOutputCallback(s.cString());
+    zfzfzfLog("%s", s.cString());
 }
 
 void zfCoreCriticalErrorCallbackPrepareAdd(ZF_IN ZFCoreCriticalErrorCallback callback)
@@ -112,6 +116,7 @@ void _ZFP_zfCoreCriticalErrorPrepare(ZF_IN const ZFCallerInfo &callerInfo)
 }
 void _ZFP_zfCoreCriticalError(ZF_IN const ZFCallerInfo &callerInfo)
 {
+    zfzfzfLog("%s", callerInfo.objectInfo().cString());
     const ZFCoreArrayPOD<ZFCoreCriticalErrorCallback> &criticalErrorCallbacks = _ZFP_ZFCoreCriticalErrorCallbacks;
     for(zfindex i = 0; i < criticalErrorCallbacks.count(); ++i)
     {