+++ ZF/ZFCore/zfsrc/ZFCore/ZFObjectDef/ZFObjectHolder.h
@@ -83,6 +83,21 @@ public:
     virtual ZFCompareResult objectCompare(ZF_IN ZFObject *anotherObj);
 
 public:
+    // override for performance
+    /** @cond ZFPrivateDoc */
+    zfoverride
+    virtual inline ZFAny objectHolded(void)
+    {
+        return this->_ZFP_objectHolded;
+    }
+    template<typename T_ZFObject>
+    T_ZFObject objectHolded(void)
+    {
+        return this->_ZFP_objectHolded;
+    }
+    /** @endcond */
+
+public:
     ZFAny _ZFP_objectHolded;
 };