+++ ZF/ZFUIKit_impl/zfsrc/ZFImpl/sys_Android/ZFProtocolZFUISysWindowEmbedNativeView_sys_Android.cpp
@@ -1,0 +1,82 @@
+/* ====================================================================== *
+ * Copyright (c) 2010-2018 ZFFramework
+ * Github repo: https://github.com/ZFFramework/ZFFramework
+ * Home page: http://ZFFramework.com
+ * Blog: http://zsaber.com
+ * Contact: master@zsaber.com (Chinese and English only)
+ * Distributed under MIT license:
+ *   https://github.com/ZFFramework/ZFFramework/blob/master/LICENSE
+ * ====================================================================== */
+#include "ZFImpl_sys_Android_ZFUIKit_impl.h"
+#include "ZFUIKit/protocol/ZFProtocolZFUISysWindowEmbedNativeView.h"
+#include "ZFUIKit/protocol/ZFProtocolZFUIView.h"
+
+#if ZF_ENV_sys_Android
+
+ZF_NAMESPACE_GLOBAL_BEGIN
+
+#define ZFImpl_sys_Android_JNI_ID_ZFUISysWindowEmbedNativeView ZFImpl_sys_Android_JNI_ID(ZFUIKit_1impl_ZFUISysWindowEmbedNativeView)
+#define ZFImpl_sys_Android_JNI_NAME_ZFUISysWindowEmbedNativeView ZFImpl_sys_Android_JNI_NAME(ZFUIKit_impl.ZFUISysWindowEmbedNativeView)
+
+ZFPROTOCOL_IMPLEMENTATION_BEGIN(ZFUISysWindowEmbedNativeViewImpl_sys_Android, ZFUISysWindowEmbedNativeView, ZFProtocolLevel::e_SystemNormal)
+    ZFPROTOCOL_IMPLEMENTATION_PLATFORM_DEPENDENCY_BEGIN()
+    ZFPROTOCOL_IMPLEMENTATION_PLATFORM_DEPENDENCY_ITEM(ZFUIView, zfText("Android:View"))
+    ZFPROTOCOL_IMPLEMENTATION_PLATFORM_DEPENDENCY_END()
+
+public:
+    zfoverride
+    virtual void protocolOnInit(void)
+    {
+        zfsuper::protocolOnInit();
+        JNIEnv *jniEnv = JNIGetJNIEnv();
+        jobject tmp = zfnull;
+
+        tmp = JNIUtilFindClass(jniEnv, JNIConvertClassNameForFindClass(ZFImpl_sys_Android_JNI_NAME_ZFUISysWindowEmbedNativeView).c_str());
+        this->jclsOwner = (jclass)JNIUtilNewGlobalRef(jniEnv, tmp);
+        JNIUtilDeleteLocalRef(jniEnv, tmp);
+    }
+    zfoverride
+    virtual void protocolOnDealloc(void)
+    {
+        JNIEnv *jniEnv = JNIGetJNIEnv();
+        JNIUtilDeleteGlobalRef(jniEnv, this->jclsOwner);
+        zfsuper::protocolOnDealloc();
+    }
+
+public:
+    virtual void nativeViewAdd(ZF_IN void *parent, ZF_IN void *child)
+    {
+        JNIEnv *jniEnv = JNIGetJNIEnv();
+        static jmethodID jmId = JNIUtilGetStaticMethodID(jniEnv, this->jclsOwner, zfTextA("native_nativeViewAdd"),
+            JNIGetMethodSig(JNIType::S_void, JNIParamTypeContainer()
+                .add(JNIType::S_object(ZFImpl_sys_Android_JNI_NAME_Object))
+                .add(JNIType::S_object(ZFImpl_sys_Android_JNI_NAME_Object))
+            ).c_str());
+        JNIUtilCallStaticVoidMethod(jniEnv, this->jclsOwner, jmId,
+            ZFCastStatic(jobject, parent),
+            ZFCastStatic(jobject, child)
+            );
+    }
+    virtual void nativeViewRemote(ZF_IN void *parent, ZF_IN void *child)
+    {
+        JNIEnv *jniEnv = JNIGetJNIEnv();
+        static jmethodID jmId = JNIUtilGetStaticMethodID(jniEnv, this->jclsOwner, zfTextA("native_nativeViewRemove"),
+            JNIGetMethodSig(JNIType::S_void, JNIParamTypeContainer()
+                .add(JNIType::S_object(ZFImpl_sys_Android_JNI_NAME_Object))
+                .add(JNIType::S_object(ZFImpl_sys_Android_JNI_NAME_Object))
+            ).c_str());
+        JNIUtilCallStaticVoidMethod(jniEnv, this->jclsOwner, jmId,
+            ZFCastStatic(jobject, parent),
+            ZFCastStatic(jobject, child)
+            );
+    }
+
+private:
+    jclass jclsOwner;
+ZFPROTOCOL_IMPLEMENTATION_END(ZFUISysWindowEmbedNativeViewImpl_sys_Android)
+ZFPROTOCOL_IMPLEMENTATION_REGISTER(ZFUISysWindowEmbedNativeViewImpl_sys_Android)
+
+ZF_NAMESPACE_GLOBAL_END
+
+#endif // #if ZF_ENV_sys_Android
+